m2 - multiple version unpublish issue fixed

+++ components/org.wso2.carbon.appmgt.impl/src/main/java/org/wso2/carbon/appmgt/impl/APIGatewayManager.java
@@ -164,8 +164,10 @@ public class APIGatewayManager {
 									  environment.getName());
 				}
 				String operation = "delete";
-				client.deleteNonVersionedWebApp(tenantDomain);
-				client.deleteVersionedWebApp(tenantDomain);
+                if (api.isDefaultVersion()) {
+                    client.deleteNonVersionedWebApp(tenantDomain);
+                }
+                client.deleteVersionedWebApp(tenantDomain);
 				undeployCustomSequences(api, tenantDomain, environment);
 			}
 		}
